refactor: extract model type to its own file
