+++ app/core/recognition/model/index.ts
@@ -1,4 +1,5 @@
 import { getGPUDelegate } from "~/core/recognition/model/delegates/gpu";
+import { ModelType } from "~/core/recognition/model/model-type";
 import Interpreter = org.tensorflow.lite.Interpreter;
 import MetadataExtractor = org.tensorflow.lite.support.metadata.MetadataExtractor;
 type Labels = string[];
@@ -128,8 +129,4 @@ export interface ModelInfo {
   name: string;
   version: string;
   author: string;
-}
-
-export enum ModelType {
-  MLP = "MLP", CNN = "CNN"
 }
