+++ app/core/recognition/model/model-type.ts
@@ -1,0 +1,14 @@
+export enum ModelType {
+  MLP = "MLP", CNN = "CNN"
+}
+
+export function modelTypeFrom(value: string): ModelType {
+  switch (value.toUpperCase()) {
+    case ModelType.MLP:
+      return ModelType.MLP;
+    case ModelType.CNN:
+      return ModelType.CNN;
+    default:
+      throw new Error(`Unsupported architecute: ${value}`);
+  }
+}
